+++ sv-parser-pp/testcases/test14.sv
@@ -1,0 +1,14 @@
+module A;
+`ifdef OPT_1
+  //wire a = 1'b1;
+`else
+  wire a = 1'b0;
+`endif
+`ifdef DEBUG
+  `ifdef OPT_2
+  //wire b = 1'b1;
+  `else
+  wire b = 1'b0;
+  `endif
+`endif
+endmodule